Reset loop file to specified loop (instead only loop 0), resolved #90

Closes #90

See merge request hi-dkfz/iml/nnactive!22

+++ nnactive/cli/subcommands/reset_loops.py
@@ -3,6 +3,7 @@ import shutil
 from argparse import Namespace
 
 from nnactive.cli.registry import register_subcommand
+from nnactive.cli.subcommands.nnunet_preprocess import preprocess
 from nnactive.cli.subcommands.update_data import update_step
 from nnactive.config import ActiveConfig
 from nnactive.loops.loading import get_sorted_loop_files
@@ -11,8 +12,10 @@ from nnactive.results.state import State
 from nnactive.results.utils import get_results_folder as get_nnactive_results_folder
 
 """
-Resets the experiment with the given dataset id to loop 0.
-Attention: This will delete all loop files from the already performed loops!
+Resets the experiment with the given dataset id to the given loop number (0 by default).
+Attention: This will delete all loop files from the already performed loops until the specified loop!
+If reset to loop 0, no preprocessing is performed to completely reset the dataset.
+Otherwise, the preprocessing is performed with the do_all flag to do preprocessing until that loop.
 """
 
 
@@ -20,41 +23,57 @@ Attention: This will delete all loop files from the already performed loops!
     "reset_loops",
     [
         (("-d", "--dataset_id"), {"type": int, "required": True}),
+        (
+            "--loop_nr",
+            {
+                "type": int,
+                "default": 0,
+                "help": "Loop number to reset to. Default is 0.",
+            },
+        ),
     ],
 )
 def main(args: Namespace) -> None:
     dataset_id = args.dataset_id
+    reset_loop_nr = args.loop_nr
     config = ActiveConfig.get_from_id(dataset_id)
     raw_dataset_path = get_raw_path(dataset_id)
-    loop_val = len(get_sorted_loop_files(raw_dataset_path))
     state = State.get_id_state(dataset_id)
-    if os.path.isfile(raw_dataset_path / f"{config.uncertainty}_{loop_val:03d}.json"):
-        print(
-            "Deleting file ",
-            str(raw_dataset_path / f"{config.uncertainty}_{loop_val:03d}.json"),
+    if reset_loop_nr > state.loop:
+        raise AttributeError(
+            "Loop number to reset to is higher than the current loop number."
         )
-        os.remove(raw_dataset_path / f"{config.uncertainty}_{loop_val:03d}.json")
-    elif os.path.isfile(
-        raw_dataset_path / f"{config.uncertainty}_{state.loop:03d}.json"
-    ):
-        print(
-            "Deleting file ",
-            str(raw_dataset_path / f"{config.uncertainty}_{state.loop:03d}.json"),
-        )
-        os.remove(raw_dataset_path / f"{config.uncertainty}_{state.loop:03d}.json")
 
-    update_step(dataset_id, loop_val=0, annotated=True, force=True)
+    for file in os.listdir(raw_dataset_path):
+        if file.startswith(f"{config.uncertainty}_") and file.endswith(".json"):
+            if int(file.split("_")[-1].split(".")[0]) > reset_loop_nr:
+                print("Deleting file ", str(raw_dataset_path / file))
+                os.remove(raw_dataset_path / file)
 
-    for loop_file in get_sorted_loop_files(raw_dataset_path)[1:]:
+    update_step(dataset_id, loop_val=reset_loop_nr, annotated=True, force=True)
+
+    for loop_file in get_sorted_loop_files(raw_dataset_path)[reset_loop_nr + 1 :]:
         print("Deleting file ", str(raw_dataset_path / loop_file))
         os.remove(raw_dataset_path / loop_file)
 
     nnactive_results_folder = get_nnactive_results_folder(dataset_id)
     for folder in os.listdir(nnactive_results_folder):
         if folder.startswith("loop"):
-            print("Deleting folder ", str(nnactive_results_folder / folder))
-            shutil.rmtree(nnactive_results_folder / folder)
+            if int(folder.split("_")[-1]) >= reset_loop_nr:
+                print("Deleting folder ", str(nnactive_results_folder / folder))
+                shutil.rmtree(nnactive_results_folder / folder)
 
     print("Resetting State file...")
     state.reset()
+    if reset_loop_nr > 0:
+        state.loop = reset_loop_nr
     state.save_state()
+
+    # Preprocess with do_all is esp. needed for loop larger 0
+    if reset_loop_nr > 0:
+        preprocess(
+            [dataset_id],
+            configurations=[config.model_config],
+            num_processes=[config.num_processes],
+            do_all=True,
+        )
